srcs/backend/requirements.txt
Django==4.2.16
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt
django-filter==24.3
Pygments==2.18.0
django-cors-headers==4.6.0
django-redis==5.4.0
channels==3.0.5
channels-redis
PyJWT